// src/shared/api/type.ts
export enum APIMethod {
    GET = 'GET',
    PUT = 'PUT',
    POST = 'POST',
    DELETE = 'DELETE',
    OPTION = 'OPTION',
  }
  
  export enum DateFormat {
    SERVER_FORMAT_DATE_TIME = 'YYYY-MM-DD HH:mm:ss',
    SERVER_FORMAT_DEFAULT = 'YYYY-MM-DD',
    CLIENT_FORMAT = 'DD/MM/YYYY',
    CLIENT_FORMAT_DATE_TIME = 'DD/MM/YYYY HH:mm:ss',
    CLIENT_FORMAT_TIME = 'HH:ss',
  }
  export enum FormType {
    IMPORT,
    EXPORT,
    ALL,
  }